make the site selector a component and make it sticky
