+++ static/client/components/MainLayout/MainLayout.tsx
@@ -1,5 +1,3 @@
-import { useCallback } from "react";
-
 import { useLocation } from "react-router-dom";
 
 import Navigation from "@/components/Navigation";
@@ -12,10 +10,6 @@ interface IMainLayoutProps {
 const MainLayout = ({ children }: IMainLayoutProps): JSX.Element => {
   const location = useLocation();
 
-  const logout = useCallback(() => {
-    window.open("/logout", "_self");
-  }, []);
-
   return (
     <>
       <div className="l-application">
@@ -26,7 +20,6 @@ const MainLayout = ({ children }: IMainLayoutProps): JSX.Element => {
             <>
               <h2>Welcome to the Content System</h2>
               <h4>Please select a page that you are looking for from the left sidebar</h4>
-              <button onClick={logout}>Logout</button>
             </>
           )}
           {children}
