+++ static/client/components/Navigation/Navigation.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useCallback, useState } from "react";
 
 import { Button } from "@canonical/react-components";
 import classNames from "classnames";
@@ -7,9 +7,14 @@ import NavigationBanner from "./NavigationBanner";
 import NavigationItems from "./NavigationItems";
 
 import NavigationCollapseToggle from "@/components/Navigation/NavigationCollapseToggle";
+import SiteSelector from "@/components/SiteSelector";
 
 const Navigation = (): JSX.Element => {
   const [isCollapsed, setIsCollapsed] = useState(true);
+
+  const logout = useCallback(() => {
+    window.open("/logout", "_self");
+  }, []);
 
   return (
     <>
@@ -33,9 +38,16 @@ const Navigation = (): JSX.Element => {
               <div className="l-navigation__controls">
                 <NavigationCollapseToggle isCollapsed={isCollapsed} setIsCollapsed={setIsCollapsed} />
               </div>
+              <SiteSelector />
             </div>
             <div className="p-panel__content">
               <NavigationItems />
+            </div>
+            <div className="p-panel__footer">
+              <Button appearance="base" onClick={logout}>
+                <i className="p-icon--logout is-light" />
+                <span>&nbsp;&nbsp;Log out</span>
+              </Button>
             </div>
           </div>
         </div>
