+++ static/client/store/types.ts
@@ -1,4 +1,8 @@
+import { type IPagesResponse } from "@/services/api/types/pages";
+
 export interface IStore {
+  selectedProject: IPagesResponse["data"] | null;
   user: string | null;
+  setSelectedProject: (s: IPagesResponse["data"]) => void;
   setUser: (u: string) => void;
 }
